pyproject.toml
[tool.poetry]
name = "transasm"
version = "0.1.0"
description = ""
authors = ["Charles Paulet <valkheim@protonmail.com>"]

[tool.poetry.scripts]
transasm = "transasm.__main__:main"
test = "transasm.scripts.test:test_unittest"
coverage = "transasm.scripts.test:test_coverage"

[tool.poetry.dependencies]
python = "^3.10"
keystone-engine = "^0.9.2"
capstone = "^5.0.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
coverage = "^6.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
